ENGR00111697 fix mx27 building failure on .28 kernel
It's intended to just fix the building error, currently we do not commit
the mx27 kernel can work on ads board. The reason is as follows. There're
two branchs for mx27ads in kernel now, one is our specific version and
the other is the mainline. Currently we do not have any plan to maintain
mx27 kernel, so we will not dig into it further before we have any update
plan(and moreover, the decision of which one we will use finally).

Signed-off-by: Robby Cai <r63905@freescale.com>