ENGR00108842-2 : MX51: Add support for CPU freq at 1GHz.
Added 1GHz CPU working point support.
Set the CPU working points based on the frequency the CPU is booted which
can be changed using the clock command in redboot.
Also fix the CPUFREQ driver so that the BogoMips are correct when CPU freq
is set to a different value in redboot.

Signed-off-by: Ranjani Vaidyanathan-RA5478 <Ranjani.Vaidyanathan@freescale.com>