ENGR00125251 usb_gadget: should not read PORT Reset at reset_irq
1. At reset_irq, the status of port reset is unsure, maybe the reset
process(hardware does it) is finished, and the status of port reset
is also cleared by usb controller. So it only needs to compare to USBSTS
at usb irq process.

2. Due to mx35/mx25 phy's bug, it needs to reset phy when re-open
usb clock next time(Begin to use usb next time)

3. mdelay 100 seconds is too long for resume process, as this code
is only added for mx37, add arch macro for this mdelay. This can
minimize the effect for other platforms.

4. Compile is ok for all imx platform, functional tests are finished
for mx35 and mx23.

Signed-off-by: Peter Chen <peter.chen@freescale.com>