ENGR00136909: MX50: Drop DDR freq to 133Mhz when AHB is at medium setpoint.
Change the DDR freq to 133MHz from 266Mhz (or 200MHz) when the AHB
is dropped to 66.5MHz. The DDR freq change will be initiated only
when the EPDC clock is not active. So there will be brief periods of time
when DDR is at 266Mhz even when AHB is at 66.5Mhz and DDR will be at 133Mhz
even when AHB is at 133Mhz.

Signed-off-by: Ranjani Vaidyanathan <ra5478@freescale.com>