ENGR00141396 ipuv3 fb: add event notification into android earlysuspend callback
As some framebuffer drivers like tve, ldb using fb notification
to get blank/unblank event, and shutdown/open the related clock.
From android view, we do not explicitly call the IOBLANK ioctl
from user space, but do it in the ipuv3 fb driver's earlysuspend.
So we must done the event notification in earlysuspend callback
instead of doing it in the IOBLANK ioctl.

Signed-off-by: Xinyu Chen <xinyu.chen@freescale.com>