ENGR00151858-1 IPUv3:Improve power stablility
This patch disables MEM_FG_SYNC channel when entering IPU
suspend. Adding this operation to IPU suspend can make sure
MEM_FG_SYNC channel is disabled, because we will not rely
on IPUv3 framebuffer driver to blank the framebuffer which
is related to MEM_FG_SYNC channel when entering suspend
mode. Note that this patch is for Android.

Signed-off-by: Liu Ying <Ying.Liu@freescale.com>