ENGR00151858-2 IPUv3 FB:Improve power stability
This patch makes IPUv3 FB driver don't blank the
framebuffer which is related with MEM_FG_SYNC channel
when entering suspend mode, but leaves the channel
disabling operation to IPUv3 driver. This makes
the overlay framebuffer won't be accidently
unblanked by up level code again, which causes
suspend or resume fail. Note that this patch is
for Android.

Signed-off-by: Liu Ying <Ying.Liu@freescale.com>