ENGR00163366 defconfig: remove CGROUP to workaround uart issue
The wierd uart issue is that: when your system partition is
mounted from SD card, the UART1 rx interrupt can not come
anymore, but if mounted root as NFS, it's ok.
There're workarounds: disable CGROUP or enable WAIT MODE.
But as WAIT MODE is not stable at all, we have to
disable CGROUP. It looks like related to cpu idle.

Signed-off-by: Xinyu Chen <xinyu.chen@freescale.com>