ENGR00175700 System hang when change HDMI from XGA to 1080P with display blank
Change the vide mode from XGA to 1080P when display blank,
the system will hang.

It is cause by overflow interrupt will trigger when the video mode change,
but clean the interrupt status bit depend on pixewl clock.
In blank state the pixel clock is gating so the HDMI PHY can't work.

Signed-off-by: Sandor Yu <R01008@freescale.com>