ENGR00176567 gpio_keys: Set the IRQF_NO_SUSPEND flag to wakeup button
We should not disabled GPIO irq when doing suspend.

During suspend, kernel will call suspend_device_irq()
to disable all the irqs (just set a disabled flag, no mask)
On the another side, The handle_level_irq will check
the irq's disable status to determin it should be
handled or unmasked.

If the GPIO interrupt happened just after the above
suspend_device_irq(), we enters handle_leve_irq(),
mask and ack the GPIO irq, then return directly
without handle or unmask this irq because of it's disabled.
That means the GPIO IMR bit is not set and
we can not wake up by the GPIO key button.

Signed-off-by: Xinyu Chen <xinyu.chen@freescale.com>