ENGR00178458 - WM8962 regulator constraint fix to prevent unwanted disable
SPKVDD regulator was being disabled whenever EPDC was included in the
image, because the EPD PMIC initialization code includes an invocation
of regulator_has_full_constraints().  This causes all regulators with
zero ref count to be disabled as part of a late_initcall.  To prevent
this disable (which breaks ethernet and DHCP), set regulator to
have boot_on attribute, so that it will not be disabled at end of
driver loading sequence.

Signed-off-by: Danny Nold <dannynold@freescale.com>