ENGR00227653 Revert "Delay interactive governor to speed up kernel boot"
The updated interactive governor added input
handler callbacks to capture the multitouch, touchpad
device's connect/disconnect/report events.
If using the interactive as default governor,
the interactive's start command will be triggered
when bringup each cpu cores, and the input handler
will be registered. But at this time, the
inputopen_work is not initialized until the
late_initcall. This caused if there's a touch
screen driver registered, and input handler connected
callback be called, the queue work to inputopen_work
will crash.

This reverts commit 2c1d402e9b2f802e33ccb622f41527812cac386f.

Signed-off-by: Xinyu Chen <xinyu.chen@freescale.com>