ENGR00236853-2: [MX28] - add suspend-to-ram feature
Add suspend-to-ram feature for MX28. In MX28 suspend-to-ram mode,
MX28 chip is off, only PWSITCH can resume it.  When resume,
it performs a cold boot, after bootlet runs,  it will  check the
persistent to decide whether to jump the specified DRAM address to resume.
Because it is cold boot, so all the registers values are lost and need be
restored when resume. The drivers need restore the register values when
resume from suspend-to-ram.

Signed-off-by: Grace Si <b18730@freescale.com>