ENGR00261425 mxc vout:Regularize video buffer timestamp
The video buffer timestamps set by user are not always
regular. Sometimes, the tv_usec field of a timeval is
greater than or equal to the macro USEC_PER_SEC(1000000),
which makes the driver wrongly determine if a video
buffer is expiring or not according to current timeval.
This patch regularizes the timestamps to fix the issue.

Signed-off-by: Liu Ying <Ying.Liu@freescale.com>