ENGR00263304-1 arm: IPUv3:Make ipu_enable_irq be able to return error
The callers of ipu_enable_irq() may choose to enable a sync interrupt
without calling ipu_request_irq() to assign an interrupt handler to
that interrupt beforehand. This is wrong and may cause NULL interrupt
handler being called in the IPU sync interrupt handler and finally
makes the system hang. This patch changes the return type of the
function ipu_enable_irq() from 'void' to 'int' so that the callers
may be aware of the error.

Signed-off-by: Liu Ying <Ying.Liu@freescale.com>