ENGR00278941-1: arm: imx: Fix suspend/resume unstable issue
As we need to float DDR IO when entering DSM, so those registers
we need to access after DDR IO is floated must be contained in TLB,
otherwise, the TLB update may case DDR access and lead to
system hang. To make sure these registers' address is in TLB,
we need to flush TLB first then access them manually.

Signed-off-by: Anson Huang <b20788@freescale.com>