ENGR00277886-2 mx28 FEC: L2 switch does not work in 10 Base-T network
L2 switch driver also does not work when the 2 ethernet connectors are
connecting to a half duplex and a full duplex network respectively. It is
because the driver saves the two ethernet phy link speed into the same
variable parameter. Because this parameter across the 2 networks is
different, the driver thinks thta link state is always changing and it
re-configure the PHY endlessly.

This code saves and recalls the duplex parameter for the 2 networks
separately

Signed-off-by: Peter Chan <B18700@freescale.com>