ENGR00278452 PU: fix system hang if run Audio loop test
cpufreq and pu power management will oprate the same register, it looks should
keep them operating in order not concurrent. Then use the cpufreq mutex in xPU
driver to make sure.
Another code bug is in pu power management which base regulator framework, we
set the reister directly not regulator, then there is the chance set_volatage
will miss the chance if this time voltage setting as the last time setting
before PU disable, but we have to do this, so add regulator_sync_voltage to
force sync the voltage setting. The drawback is setting twice...

Note:
This patch is temporary and supposed to be replaced it once we find the
root cause.

Signed-off-by: Robin Gong <b38343@freescale.com>
Acked-by: Jason Liu