ENGR00303539-4 video: backlight: Add fb event info in bl dev
The check_fb() hook may need the fb event information to
do something.  For example, people may use it to delay
some time before enabling a backlight as some special panel
requires.  To support this, this patch adds a field of
struct fb_event in the structure backlight_device and sets
the field in the fb notifier handler of the backlight driver.

Signed-off-by: Liu Ying <Ying.Liu@freescale.com>