ENGR00303539-6 MX6sl EVK: Add check_fb hook in PWM bl pdata
There is a SEIKO WVGA panel on the MX6sl EVK platform.
The panel's backlight should be enabled after DISP pin and
fb are active for at least 170ms(10 frames), according to
the panel's spec.  The timing could be described by the
following graph.
                 _____________________________________
power      _____|                                     |_____
                    ________________________________
DISP       ________|                                |_______
           ________  _____  _______________________  _______
           invalid \/blank\/        valid          \/
data       ________/\_____/\_______________________/\_______
                 ->| 170ms  |<-
                             ______________________
backlight  _________________|                      |________

In order to guarantee the specific timing to enable the
backlight, this patch adds a check_fb() hook in the PWM
backlight platform data.  We choose to use 200ms delay,
which should be safe enough.

Signed-off-by: Liu Ying <Ying.Liu@freescale.com>