ENGR00313685-15 ARM: dts: imx: apply ENET IRQ workaround for sabresd board
This is pretty much an example to demonstrate how the GPIO6 workaround
for bug ERR006687 (ENET: Only the ENET wake-up interrupt request can
wake the system from Wait mode) should be applied for a board.

Basically it requires a setup of MX6QDL_PAD_GPIO_6__ENET_IRQ in pinctrl
entry, and an overwrite on the property interrupts-extended to replace
the ENET GIC IRQ with GPIO1_6.

Since the pad GPIO6 is used by I2C3 on the board, we have to create
sabresd-enetirq.dts with I2C3 disabled to enable this workaround.

Signed-off-by: Shawn Guo <shawn.guo@freescale.com>