ENGR00318063-7: ARM: imx6q: mmdc_ch_axi clock should be non-gateable
Per the updated CCM chapter, there is no gate for clock mmdc_ch1_axi.
And there is a gate for mmdc_ch0_axi, but the gate also controls FABRIC
clock, which means the gate cannot be gated anyway.  That said,
implementing these two clocks as gate does not make too much sense.

Since the clocks are already referenced in quite some places, to
minimize the impact of dropping the gate, instead of removing the clocks
completely, we choose to point the clocks to their dividers.

Signed-off-by: Shawn Guo <shawn.guo@freescale.com>