ENGR00331168 ARM: IMX6: lcdif pix clk's rate should be set under the clk is disabled
One of the ancestors is 'pll5_video', and a previous commit
"ARM: imx: pllv3 can only be configured when it's powered off"
has added power off check when set a pll clock's rate. So when
the pix clk's rate is set, it should not be enabled. Otherwise,
the kernel will print below error message:

"clk_pllv3_av_set_rate: cannot configure divider when PLL is powered on".

Signed-off-by: Fancy Fang <chen.fang@freescale.com>