ENGR00327364 [imx6x] Ensure that the bandgap self-bias circuit is disabled after boot.
Back ported from 3.10.x kernel:
516bf00a8ce530a269f6d2f1470f88e299584524

The self-bias circuit is used by the bandgap during startup.
Once the bandgap has stabilized, the self-bias circuit should be
disabled for best noise performance of analog blocks.
Also this bit should be disabled before the chip enters STOP mode or
when ever the regular bandgap is disabled.
This patch ensures that the self-bais bit is cleared before the bandgap
is disabled in low power code and during suspend/resumne.

Signed-off-by: Ranjani Vaidyanathan <Ranjani.Vaidyanathan@freescale.com>