MLK-9742-5 usb: common: otg-fsm: schedule HNP polling even the transfer has failed
At current code, if the host fails to get device's request flag, it returns
-EIO, and will not restart the HNP polling timer, it causes the HNP polling
does not work after that, we will see "usb 1-1: Get one byte OTG status failed"
for this error case. We may meet above case when the bus is not active (No SOF),
for example: during the bus suspend or the bus reset process.

Signed-off-by: Peter Chen <peter.chen@freescale.com>