MLK-9856: ASoC: imx-hdmi: fix the failure in module test
The kernel log is

[<8000dcd8>] (__dabt_svc+0x38/0x60) from [<8067d218>] (mutex_lock+0xc/0x4c)
[<8067d218>] (mutex_lock+0xc/0x4c) from [<804f8c48>] (snd_soc_dapm_stream_event+0x20/0xe4)
[<804f8c48>] (snd_soc_dapm_stream_event+0x20/0xe4) from [<804fa5b4>] (close_delayed_work+0x3c/0x48)
[<804fa5b4>] (close_delayed_work+0x3c/0x48) from [<80042c68>] (process_one_work+0xfc/0x34c)
[<80042c68>] (process_one_work+0xfc/0x34c) from [<80043700>] (worker_thread+0x144/0x3a4)
[<80043700>] (worker_thread+0x144/0x3a4) from [<80048d7c>] (kthread+0xa4/0xb0)
[<80048d7c>] (kthread+0xa4/0xb0) from [<8000e1d8>] (ret_from_fork+0x14/0x3c)

The alsa driver use shedule_delayed_work when close pcm in remove module.
But after remove module the resource is released, so there will be kernel dump.
So here use ignore_pmdown_time to avoid to use shedule_delayed_work, then this
issue can be avoided.

Signed-off-by: Shengjiu Wang <shengjiu.wang@freescale.com>