MLK-10057 PxP V4L2 output: force to call pxp_streamoff when device closes
By previous implementation, there's the possibility that last picture remains
on screen when the program exits. This can be reproduced by not calling
STREAMOFF ioctl in v4l2 output application or just trying to kill the v4l2
output program. The driver has faults to handle this case, since it depends on
'pxp->s0_vbq.streaming' variable be true in close() function to call
pxp_streamoff() while the variable is set to 0 after the application calls
munmap().  The driver should call pxp_streamoff() even if the application
does not call STREAMOFF ioctl.

This patch uses the local 'streaming' variable to track the streaming
status to fix this problem.

Signed-off-by: Robby Cai <r63905@freescale.com>