MLK-10210 WIFI: ath6kl: Fixed the issue that warning messages were printed when uninstalling ath6kl_sdio after iperf test.
When handing the command "ifconfig wlan0 down", the driver handler did
not notify the disconnected event to the upper layer stack cfg80211,
which would cause the relating resources unable to be released. A
notification of disconnet event to cfg80211 was added when the driver
was changed to the disconnect state.

Signed-off-by: Shenwei Wang <shenwei.wang@freescale.com>
(cherry picked from commit 0b7b203a563ac98f5f0c5695a5149ec80d2114b4)