MLK-13996: lcdif: Use DE polarity specified in DTS
Currently, the LCDIF driver (mxsfb.c) is overriding the DE polarity
specified in DTS with the one specified in fb_videomode (sync member) by
the panel driver.
Initially, the panel driver found in
drivers/video/fbdev/mxc/mxcfb_hx8363_wvga.c specified the sync in
fb_videomode as FB_SYNC_OE_LOW_ACT.

But this patch, changed it to 0x0:

commit 4deb430fd05a ("MLK-13607-8 video: mipi-panel: hx8363: change DE
polarity to active high")
Author: Fancy Fang <chen.fang@nxp.com>
Date:   Wed Dec 14 16:21:51 2016 +0800

    The NorthWest Logic MIPI DSI only support active low data enable
    signal on imx7ulp-evk board.

The problem is that, the hx8363 panel was working with active low on
imx7d-sdb, but with active high on imx7ulp-evk. The above patch broke
hx8363 panel on imx7d-sdb board.

So, instead of using a hard-coded polarity in panel driver, better use
the one defined in device-tree.

Signed-off-by: Robert Chiras <robert.chiras@nxp.com>