MLK-14542 ARM: dts: imx7d-sdb: change DE polarity to high for HDMI and LCD
Fix HDMI functionality by changing DE default polarity to active high,
which is needed by the default HDMI mode. Active low DE is needed by MIPI
DSI, which is why there is a different dts file for working with a MIPI
DSI panel.

This issue has been previously fixed in commit 5443a75ed03 ("MLK-14283: dts:
fix DE polarity for lcdif"), but this commit was reverted because it broke
other workflows. It was reverted in commit 34ac60798ec ("Revert "MLK-14283:
dts: fix DE polarity for lcdif"")and another solution was offered in
commit 8766ca8eddf ("MLK-14399: 4.9 rebase: LVDS panel does not work on
iMX6SX Auto").

This change of the default DE polarity was one of the changes in the
reverted commit, which should have been kept.

Signed-off-by: Cristina Ciocan <cristina-mihaela.ciocan@nxp.com>