MLK-16108 PCI: imx: turn on pd for imx8mq pcie
Root cause:
Poewr domain of the PCIEs are turned off, and
not turned on properly in previous ATF.

The PDs of PCIE1/2 have the dependency.
Both of the PDs should be operated at same time.
This issue is gone after update the PDs operations
in ATF.
In order to make sure that the PDs are turned on,
Turn power domain for imx8mq pcie explicitly in
driver.

Signed-off-by: Richard Zhu <hongxing.zhu@nxp.com>