MLK-15985-1: ASoC: fsl_ssi: remove the wrong fix for S20_3LE
This reverts commit 6ce4e9c184b7 ("MLK-15068: ASoC: fsl_ssi: fix the noise
issue with S20_3LE Mono bitsream")

The fix in MLK-15068 can't fix the mono noise issue, for using the
physical width imply that the sample with is 24, but the
CCSR_SSI_SxCCR_WL still using the 20 bit, the unalignment cause noise.
Or if change the CCSR_SSI_SxCCR_WL to 24bit, the volume is lower
for 24bit imply that the sample is shift 4bit right.

So the correct way is to change the mclk frequency in dts.

Signed-off-by: Shengjiu Wang <shengjiu.wang@nxp.com>