MLK-17470-2: ASoC: ak4497: review get/set_dsdsel
With commit 69201427e188 ("MLK-17470: ASoC: ak4497: automatically select dsdsel
in driver") the "ak4497_priv.nDSDSel" field might not reflect the real dsdsel
option. In order to simplify the approach and avoid future inconsistencies the
field is removed and dsdsel option is read/stored directly to/from codec
registers. Aside of this a missing dsdsel option for 22.5792MHz is introduced
in "ak4497_set_dsdsel" method, now this method being fully compatible with
commit 69201427e188 ("MLK-17470: ASoC: ak4497: automatically select dsdsel in
driver").

Signed-off-by: Viorel Suman <viorel.suman@nxp.com>