MLK-17380-5 usb: host: xhci: do not return error status for URB
Current XHCI implementation does not consider completion interrupt
for SETUP packet standalone, so it will show warning message
and return error status for URB. In fact, it can support it. In
this commit, we delete warning message and set status as zero
for URB.

Support completion interrupt for SETUP packet is needed for EH
SINGLE_STEP_SET_FEATURE Test.

Acked-by: Jun Li <jun.li@nxp.com>
Signed-off-by: Peter Chen <peter.chen@nxp.com>