MLK-17887 Revert "usb: dwc3: gadget: skip Set/Clear Halt when invalid"
This reverts commit 3ccf60e1ff5e
("usb: dwc3: gadget: skip Set/Clear Halt when invalid")
"At least macOS seems to be sending
ClearFeature(ENDPOINT_HALT) to endpoints which
aren't Halted. This makes DWC3's CLEARSTALL command
time out which causes several issues for the driver."

but this change is impacting usb CV compliance test with
windows, which using clear EP halt for other purpose in
this case, as this is not resolved on upstream, so revert
this patch for now.

Acke-by: Peter Chen <peter.chen@nxp.com>
Signed-off-by: Li Jun <jun.li@nxp.com>