MA-11050 [ATH10K]Set wifi scan results' timestamp using monotonic time
For ath10k driver, it do not set the netlink attributes for a BSS:
NL80211_BSS_LAST_SEEN_BOOTTIME, which is CLOCK_BOOTTIME timestamp
when this entry was last updated by a received frame.
But it do support NL80211_BSS_TSF which is known to be from a probe
response and NL80211_BSS_BEACON_TSF, which represents TSF of the
last received beacon.
Android does not want the timestamp from the frame. Instead it wants
a monotonic increasing value.
So we updated the timestamp with monotonic boottime.

Change-Id: I49036e51b3661ef86340bfccb37c5d99629f98f5
Signed-off-by: Tian Yang<yang.tian@nxp.com>