MLK-17733: arm64: dts: fsl-imx8qm: Fix LVDS dual-channel panel
The LVDS dual-channel pannel doesn't use the IT6263 LVDS-HDMI converter,
therefore we need to disable those nodes such that the IT6263 driver
won't disable the LDB nodes.
This is a special case, in which auto-detection cannot be used to detect
whether or not we have something connected on the LVDS ports.

Signed-off-by: Robert Chiras <robert.chiras@nxp.com>