MA-11774 [7ULP] Wifi can not scan any ap
For 7ulp, it uses 1PJ wifi module, and qcacld-2.0 driver will
update bss info with NL80211_BSS_LAST_SEEN_BOOTTIME which should
be a nanoseconds value. And for Android it should be a monotonic
increasing value instead for timestamp from frame.

qcacld-2.0 requires config cnss driver on qcom platform to enable using
monotonic time, so we revert this config for imx devices.

Change-Id: Ica65473e10ad3dc2f832f134a603724a0088df22
Signed-off-by: Tian Yang<yang.tian@nxp.com>