MLK-18239 ARM64: dts: imx8qxp: change the rpmsg reserved memory region
In order to enlarge the CMA easily, change the rpmsg
reserved memory region from 0xb800_0000 to 0x9000_0000.
And refine the layout of the reserved memory.
- RPMSG buffers are allocated from CMA dynamically, and have to
be accessed by M4 side. But M4 can only access the 1.5Gbytes
DDR memory from 0x8000_0000. So, the finial reserved memory
layout is just like the one below.

Thus, the largest size of the DDR memory left for CMA, is about
1212Mbytes in theory, since 32Mbytes alignment is required by
CMA allocation.

reserved-memory layout
0x8000_0000 ~ 0x83FF_FFFF            A core + Linux Kernel(64M)
0x8400_0000 ~ 0x85FF_FFFF            VPU encoder boot(32M)
0x8600_0000 ~ 0x87FF_FFFF            VPU decoder boot(32M)
0x8800_0000 ~ 0x8FFF_FFFF            M4 + RTOS(128M)
0x9000_0000 ~ 0x903F_FFFF            RPMSG Vring(4M)
0x9040_0000 ~ 0x913F_FFFF            VPU decoder rpc(16M)
0x9140_0000 ~ 0x923F_FFFF            VPU encoder rpc(16M)
0x9240_0000 ~ 0x943F_FFFF            DSP(32M)
0x9440_0000 ~ 0xDFFF_FFFF(MAX)       CMA(1212M)(MAX)

Signed-off-by: Richard Zhu <hongxing.zhu@nxp.com>