MLK-18272: ARM64: dts: refine the power domain tree for audio devices
In the latest scfw design, the power domain of device should be explicit
enabled in kernel, otherwise there will be kernel dump.
For example, when using audio device to playback, the DMA channel's power
domain should be eanbled, but to avoid to call scfw API in driver, we need
to refine the tree of power domain, define the DMA channel's power domain
as audio device's parent.

Signed-off-by: Shengjiu Wang <shengjiu.wang@nxp.com>